Fix parsing of show spanning-tree summary in edge cases
For configurations that only have a single VLAN or MST instance, show
spanning-tree summary prints vlan or mst in singular form. Therefore
adjust the regular expression for matching to make the trailing s
optional. This fixes parsing of the output on IOS XE with such a
configuration.

Add test coverage for those cases as well.