IOSXR: fix AS parsing for dynamic neighbors
Dynamic neighbors get an additional "Dynamic" property on the line
where remote and local AS are specified:

```
Remote AS 65498, local AS 65499, external link, Dynamic
```

Currently, the regex to parse this line expects a precise set of
properties and reject any unkown property. I propose to extend the
regex to accept and discard any unknown property. It seems better than
to return neighbors without local/remote AS, as this is quite an
essential information.