Fix 3 line output for nxos ShowBgpVrfAllAllSummary
With IPv6 addresses the show bgp summary output sometimes has to wrap
to two lines. When using a 4-byte ASN it can cause it to wrap to 3
lines, the first line for the IP, the second line for the version and
the ASN, and the third line the remaining values. This change parses
3 line output correctly.