Fix 3 line output for nxos ShowBgpVrfAllAllSummary
On NXOS if the IPv6 address is too long it will cause the other values
to wrap to a new line. If the AS on the second line is too long (in the
case of a 4-byte ASN) it will wrap to a third line. Like below:

Neighbor        V    AS MsgRcvd MsgSent   TblVer  InQ OutQ Up/Down...
2001:db8:c006:ff:251::1
                4 406942
                              0       0        0    0    0 05:49:52 Idle

Currently the parser will error out on 3-line cases, and this change
adds a fix to parse them correctly. The fix adds a p8_2_1 regex that
matches the 2nd line. In 3-line case, p8_1 will match the neighbor,
the newly added p8_2_1 will match the version and AS, and the existing
p8_4 will match the remaining values.